scripts/requirements.txt
certifi==2023.7.22
charset-normalizer==3.1.0
idna==3.4
Pillow==10.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==6.1.2
qrcode==7.4.2
requests==2.31.0
six==1.16.0
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==2.0.7
